feat: ScheduleComplete DataStore 구현

+++ app/src/main/java/pnu/plato/calendar/data/remote/repository/RemoteScheduleRepository.kt
@@ -210,7 +210,7 @@ private fun String.parseIcsToPersonalSchedules(): List<PersonalSchedule> {
 
 private fun buildScheduleFromFields(fields: Map<String, String>): PersonalSchedule =
     PersonalSchedule(
-        id = fields["UID"].orEmpty().split("@")[0].toLong(),
+        id = fields["UID"].orEmpty().split("@")[0],
         title = fields["SUMMARY"].orEmpty(),
         description = fields["DESCRIPTION"],
         startAt = fields["DTSTART"].orEmpty().parseUctToLocalDateTime(),
